management-app/requirements.txt
azure-iot-hub
six
